Removed security file, switched to jwt-extended, added jwt optional

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity
 from models.item import ItemModel
 
 items = []
@@ -18,7 +18,7 @@ class Item(Resource):
         help="This field cannot be left blank!"
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -40,14 +40,18 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
 
         return {'message': 'Item deleted'}
 
-    @jwt_required()
+    @jwt_required
     def put(self, name):
         payload = Item.parser.parse_args()
 
@@ -66,6 +70,8 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]} # Returns all of the objects in the database
+        user_id = get_jwt_identity()
+        return {'items': [item.json() for item in ItemModel.find_all()]} # Returns all of the objects in the database
